Simplify floating point numbers using %g.
We could go further by reducing `0.` to `0`, but that has the slight
potential of generating a different program (`1/2` vs `1./2.`).

sksl_binary_size_gpu: 18510 -> 18352 bytes
sksl_binary_size_graphite: 22611 -> 22526 bytes
sksl_binary_size_compute: 245 -> 245 bytes

Bug: skia:13775
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/585973
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>

+++ tests/SkSLDSLTest.cpp
@@ -192,11 +192,11 @@ DEF_GANESH_TEST_FOR_MOCK_CONTEXT(DSLFlags, r, ctxInfo) {
 DEF_GANESH_TEST_FOR_MOCK_CONTEXT(DSLFloat, r, ctxInfo) {
     AutoDSLContext context(ctxInfo.directContext()->priv().getGpu());
     Expression e1 = Float(std::numeric_limits<float>::max());
-    REPORTER_ASSERT(r, atof(e1.release()->description().c_str()) ==
+    REPORTER_ASSERT(r, std::stof(e1.release()->description().c_str()) ==
                        std::numeric_limits<float>::max());
 
     Expression e2 = Float(std::numeric_limits<float>::min());
-    REPORTER_ASSERT(r, atof(e2.release()->description().c_str()) ==
+    REPORTER_ASSERT(r, std::stof(e2.release()->description().c_str()) ==
                        std::numeric_limits<float>::min());
 
     EXPECT_EQUAL(Float2(0),
@@ -252,12 +252,12 @@ DEF_GANESH_TEST_FOR_MOCK_CONTEXT(DSLFloat, r, ctxInfo) {
 DEF_GANESH_TEST_FOR_MOCK_CONTEXT(DSLHalf, r, ctxInfo) {
     AutoDSLContext context(ctxInfo.directContext()->priv().getGpu());
     Expression e1 = Half(std::numeric_limits<float>::max());
-    REPORTER_ASSERT(r,
-                    atof(e1.release()->description().c_str()) == std::numeric_limits<float>::max());
+    REPORTER_ASSERT(r, std::stof(e1.release()->description().c_str()) ==
+                       std::numeric_limits<float>::max());
 
     Expression e2 = Half(std::numeric_limits<float>::min());
-    REPORTER_ASSERT(r,
-                    atof(e2.release()->description().c_str()) == std::numeric_limits<float>::min());
+    REPORTER_ASSERT(r, std::stof(e2.release()->description().c_str()) ==
+                       std::numeric_limits<float>::min());
 
     EXPECT_EQUAL(Half2(0),
                 "half2(0.0)");
